--- src/main/java/drug_search/frameworks_drivers/DrugSearchResultScreen.form
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="drug_search.frameworks_drivers.DrugSearchResultScreen">
-  <grid id="27dc6" binding="drugSearchResultPanel" layout-manager="GridLayoutManager" row-count="1" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="20" y="20" width="500" height="400"/>
-    </constraints>
-    <properties/>
-    <border type="none"/>
-    <children>
-      <scrollpane id="30e65" binding="scroller">
-        <constraints>
-          <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="7" hsize-policy="7" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties/>
-        <border type="none"/>
-        <children>
-          <component id="53b61" class="javax.swing.JTable" binding="drugSearchResultTable">
-            <constraints/>
-            <properties>
-              <autoCreateRowSorter value="true"/>
-            </properties>
-          </component>
-        </children>
-      </scrollpane>
-    </children>
-  </grid>
-</form>
